requirements.txt
# Pipeline (Pass 1–3)
pandas>=2.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21
openai>=1.0.0

# FastAPI API for n8n
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv